address final toolkit review comments
